refactor: moving getTradingAccountName into utils
