+++ packages/utils/src/index.ts
@@ -1,3 +1,4 @@
+export { default as getTradingAccountName } from './getTradingAccountName';
 export { default as getWalletCurrencyIcon } from './getWalletCurrencyIcon';
 export { default as groupTransactionsByDay } from './groupTransactionsByDay';
 export { default as unFormatLocaleString } from './unFormatLocaleString';
